Add _full variants to usbmuxd_{un}subscribe()
Currently, with the usbmuxd_subscribe API, it's only possible to have
one callback listening to usbmuxd events per process. This can be
problematic in case 2 different libraries (not knowing of each other)
both need access to these events.
This commit adds some _full variants which return a callback id, and
which use a collection to be able to associate multiple callbacks to be
called on usbmuxd events.